+++ social_core/tests/backends/open_id.py
@@ -1,11 +1,17 @@
 # -*- coding: utf-8 -*-
-import datetime
-import json
-import sys
 from calendar import timegm
 
-import jwt
+import os
+import sys
+import json
+import datetime
+
+from jwkest.jwk import RSAKey, KEYS
+from jwkest.jws import JWS
+from jwkest.jwt import b64encode_item
+
 import requests
+
 from openid import oidutil
 
 
@@ -26,9 +32,8 @@ from ...backends.utils import load_backends
 from ...exceptions import AuthTokenError
 from .base import BaseBackendTest
 from ..models import TestStorage, User, TestUserSocialAuth, \
-                                TestNonce, TestAssociation
+    TestNonce, TestAssociation
 from ..strategy import TestStrategy
-
 
 # Patch to remove the too-verbose output until a new version is released
 oidutil.log = lambda *args, **kwargs: None
@@ -125,7 +130,29 @@ class OpenIdConnectTestMixin(object):
     client_key = 'a-key'
     client_secret = 'a-secret-key'
     issuer = None  # id_token issuer
-    id_token_max_age = 600  # seconds
+    openid_config_body = None
+    key = None
+
+    def setUp(self):
+        super(OpenIdConnectTestMixin, self).setUp()
+        here = os.path.dirname(__file__)
+        self.key = RSAKey(kid='testkey').load(os.path.join(here, '../testkey.pem'))
+        HTTPretty.register_uri(HTTPretty.GET,
+                               self.backend.OIDC_ENDPOINT + '/.well-known/openid-configuration',
+                               status=200,
+                               body=self.openid_config_body
+                               )
+        oidc_config = json.loads(self.openid_config_body)
+
+        def jwks(_request, _uri, headers):
+            ks = KEYS()
+            ks.add(self.key.serialize())
+            return 200, headers, ks.dump_jwks()
+
+        HTTPretty.register_uri(HTTPretty.GET,
+                               oidc_config.get('jwks_uri'),
+                               status=200,
+                               body=jwks)
 
     def extra_settings(self):
         settings = super(OpenIdConnectTestMixin, self).extra_settings()
@@ -133,9 +160,7 @@ class OpenIdConnectTestMixin(object):
             'SOCIAL_AUTH_{0}_KEY'.format(self.name): self.client_key,
             'SOCIAL_AUTH_{0}_SECRET'.format(self.name): self.client_secret,
             'SOCIAL_AUTH_{0}_ID_TOKEN_DECRYPTION_KEY'.format(self.name):
-                self.client_secret,
-            'SOCIAL_AUTH_{0}_ID_TOKEN_MAX_AGE'.format(self.name):
-                self.id_token_max_age
+                self.client_secret
         })
         return settings
 
@@ -144,7 +169,7 @@ class OpenIdConnectTestMixin(object):
         Get the nonce from the request parameters, add it to the id_token, and
         return the complete response.
         """
-        nonce = parse_qs(request.body).get('nonce')
+        nonce = self.backend.data['nonce'].encode('utf-8')
         body = self.prepare_access_token_body(nonce=nonce)
         return 200, headers, body
 
@@ -166,7 +191,7 @@ class OpenIdConnectTestMixin(object):
 
         return id_token
 
-    def prepare_access_token_body(self, client_key=None, client_secret=None,
+    def prepare_access_token_body(self, client_key=None, tamper_message=False,
                                   expiration_datetime=None,
                                   issue_datetime=None, nonce=None,
                                   issuer=None):
@@ -175,15 +200,12 @@ class OpenIdConnectTestMixin(object):
 
         client_id       -- (str) OAuth ID for the client that requested
                                  authentication.
-        client_secret   -- (str) OAuth secret for the client that requested
-                                 authentication.
         expiration_time -- (datetime) Date and time after which the response
                                       should be considered invalid.
         """
 
         body = {'access_token': 'foobar', 'token_type': 'bearer'}
         client_key = client_key or self.client_key
-        client_secret = client_secret or self.client_secret
         now = datetime.datetime.utcnow()
         expiration_datetime = expiration_datetime or \
                               (now + datetime.timedelta(seconds=30))
@@ -194,21 +216,26 @@ class OpenIdConnectTestMixin(object):
             client_key, timegm(expiration_datetime.utctimetuple()),
             timegm(issue_datetime.utctimetuple()), nonce, issuer)
 
-        body['id_token'] = jwt.encode(id_token, client_secret,
-                                      algorithm='HS256').decode('utf-8')
+        body['id_token'] = JWS(id_token, jwk=self.key, alg='RS256').sign_compact()
+        if tamper_message:
+            header, msg, sig = body['id_token'].split('.')
+            id_token['sub'] = '1235'
+            msg = b64encode_item(id_token).decode('utf-8')
+            body['id_token'] = '.'.join([header, msg, sig])
+
         return json.dumps(body)
 
-    def authtoken_raised(self, expected_message_regexp, **access_token_kwargs):
+    def authtoken_raised(self, expected_message, **access_token_kwargs):
         self.access_token_body = self.prepare_access_token_body(
             **access_token_kwargs
         )
-        with self.assertRaisesRegexp(AuthTokenError, expected_message_regexp):
+        with self.assertRaisesRegexp(AuthTokenError, expected_message):
             self.do_login()
 
-    def test_invalid_secret(self):
+    def test_invalid_signature(self):
         self.authtoken_raised(
             'Token error: Signature verification failed',
-            client_secret='wrong!'
+            tamper_message=True
         )
 
     def test_expired_signature(self):
@@ -226,10 +253,10 @@ class OpenIdConnectTestMixin(object):
                               client_key='someone-else')
 
     def test_invalid_issue_time(self):
-        issue_datetime = datetime.datetime.utcnow() - \
-                         datetime.timedelta(seconds=self.id_token_max_age + 1)
+        expiration_datetime = datetime.datetime.utcnow() - \
+                              datetime.timedelta(hours=1)
         self.authtoken_raised('Token error: Incorrect id_token: iat',
-                              issue_datetime=issue_datetime)
+                              issue_datetime=expiration_datetime)
 
     def test_invalid_nonce(self):
         self.authtoken_raised(